Add vertical tabs to modal
